final commit
I didn't get things integrated like I wanted to but at this point its
the best that I can do - I'll keep working on my javascript and jquery
skills this is where I've consistently been getting hung up and as a
result my CSS suffers.  Its not because of lack of trying I've put in
countless hours and will continue to do so.